+++ src/app/components/profile/profile.module.ts
@@ -7,7 +7,7 @@ import { ProfileEditComponent } from './profile-edit/profile-edit.component';
 import { ProfileRoutingModule } from './profile-routing.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { ButtonComponent } from '@app/shared/components/button/button.component';
-import { DeleteNotificationComponent } from './profile-edit/delete-notification/delete-notification.component';
+import { DeleteNotificationComponent } from '../../modals/delete-notification/delete-notification.component';
 
 @NgModule({
   declarations: [
