+++ src/app/components/quiz-shows/quiz-shows.module.ts
@@ -7,7 +7,7 @@ import { PresentAsTitlePipe } from '@app/shared/pipes/present-as-title.pipe';
 import { ButtonComponent } from '@app/shared/components/button/button.component';
 import { QuizQuestionComponent } from './quiz-question/quiz-question.component';
 import { QuizTimerComponent } from './components/quiz-timer/quiz-timer.component';
-import { AchievementNotificationComponent } from './components/achievement-notification/achievement-notification.component';
+import { AchievementNotificationComponent } from '../../modals/achievement-notification/achievement-notification.component';
