+++ src/app/modals/achievement-notification/achievement-notification.component.scss
