+++ src/app/modals/delete-notification/delete-notification.component.scss
